requirements.txt
pydantic==1.10.13
chromadb==0.3.29
fastapi==0.85.1
langchain==0.0.350
langchain-core==0.1.17
langchain-community==0.3.7
pydantic-settings==2.0.0
openai
streamlit
python-docx
pandas
requests
python-dotenv
pysqlite3-binary
unstructured